services/shipments/requirements.txt
fastapi
prometheus_fastapi_instrumentator
uvicorn
psycopg2
pydantic
typing
python-dotenv
python-multipart
sqlalchemy